+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 News
 ====
 
-3.0.1 (unreleased)
+3.0.1 (2024-08-30)
 ------------------
 
 - Multiple file input support.